lab tech grammar changes
Fixing awkward grammar - under the first point in lab techs, one doesn't "help foster an atmosphere of a positive work environment" so I took the last part out (it's still in the other sections, though) as I think it's conveyed in the other descriptions. Last point in lab techs is a little awkward but I changed it to make it consistently start with a verb as all the others do (aside from the first, which starts "in expected to," which is then implied in all the following points...). Also, check consistency between talking in second person (you) and third person (they).